CS443-BLENDSv2/requirements.txt
jupyter==1.0.0
requests==2.20
Flask==1.0.2
sqlalchemy==1.2.14
pycryptodome==3.7.2